Always try to build prebuilt 4c docker image
The prebuilt 4C docker image is always built and tagged as `main` even
if the tests fail. This better reflects the tag name `main` as the
docker image reflects the current state on main.
If the nightly tests succeed this image gets the tag `stable`. So, the
tag `stable` reflects the last state where the tests where successful.
See issue #1134.